updated for version 7.2.383 Problem:    Vim doesn't build cleanly with MSVC 2010. Solution:   Change a few types. (George Reilly)

+++ src/if_python.c
@@ -2080,7 +2080,7 @@ WindowSetattr(PyObject *self, char *name, PyObject *val)
 	    return -1;
 
 	/* When column is out of range silently correct it. */
-	len = STRLEN(ml_get_buf(this->win->w_buffer, lnum, FALSE));
+	len = (long)STRLEN(ml_get_buf(this->win->w_buffer, lnum, FALSE));
 	if (col > len)
 	    col = len;
